Move the storyless header component down from app level and document it
The header component ia refered to from the search bar component. This
meant refering to something from a component level upwards to app level,
and was thus moved down from the app level into the search bar component
folder, where it is called.
